feat: broke up tests into multiple test files closes #38
